bitcoin_coordinator::coordinator
Trait BitcoinCoordinatorApi

pub trait BitcoinCoordinatorApi {
    // Required methods
    fn is_ready(&self) -> Result<bool, BitcoinCoordinatorError>;
    fn tick(&self) -> Result<(), BitcoinCoordinatorError>;
    fn monitor(
        &self,
        data: TypesToMonitor,
    ) -> Result<(), BitcoinCoordinatorError>;
    fn dispatch(
        &self,
        tx: Transaction,
        speedup: Option<SpeedupData>,
        context: String,
        block_height: Option<BlockHeight>,
    ) -> Result<(), BitcoinCoordinatorError>;
    fn cancel(
        &self,
        data: TypesToMonitor,
    ) -> Result<(), BitcoinCoordinatorError>;
    fn add_funding(&self, utxo: Utxo) -> Result<(), BitcoinCoordinatorError>;
    fn get_transaction(
        &self,
        txid: Txid,
    ) -> Result<TransactionStatus, BitcoinCoordinatorError>;
    fn get_news(&self) -> Result<News, BitcoinCoordinatorError>;
    fn ack_news(&self, news: AckNews) -> Result<(), BitcoinCoordinatorError>;
}

Required Methods

fn is_ready(&self) -> Result<bool, BitcoinCoordinatorError>

Checks if the coordinator is ready to process transactions Returns true if the coordinator is ready, false otherwise

fn tick(&self) -> Result<(), BitcoinCoordinatorError>

Processes pending transactions and updates their status This method should be called periodically to keep the coordinator state up-to-date

fn monitor(&self, data: TypesToMonitor) -> Result<(), BitcoinCoordinatorError>

Registers a type of data to be monitored by the coordinator The data will be tracked for confirmations and status changes, and updates will be reported through the news.

Arguments

data - The data to monitor

fn dispatch( &self, tx: Transaction, speedup: Option<SpeedupData>, context: String, block_height: Option<BlockHeight>, ) -> Result<(), BitcoinCoordinatorError>

Dispatches a transaction to the Bitcoin network

Arguments

tx - The Bitcoin transaction to dispatch

speedup - Speed up information for the transaction (None means it should not be speed up)

context - Additional context information for the transaction to be returned in news

block_height - Block height to dispatch the transaction (None means now)

fn cancel(&self, data: TypesToMonitor) -> Result<(), BitcoinCoordinatorError>

Cancels the monitor and the dispatch of a type of data This method removes the monitor and the dispatch from the coordinator’s store. Which means that the data will no longer be monitored.

Arguments

data - The data to cancel

fn add_funding(&self, utxo: Utxo) -> Result<(), BitcoinCoordinatorError>

Registers funding information for potential transaction speed-ups This allows the coordinator to create child pays for parents transactions when needed

Arguments

utxo - Utxo to use for speed-ups

fn get_transaction( &self, txid: Txid, ) -> Result<TransactionStatus, BitcoinCoordinatorError>

fn get_news(&self) -> Result<News, BitcoinCoordinatorError>

Retrieves news about monitored transactions Returns information about transaction confirmations.

fn ack_news(&self, news: AckNews) -> Result<(), BitcoinCoordinatorError>

Acknowledges that news has been processed This prevents the same news from being returned in subsequent calls to get_news()

Arguments

news - The news items to acknowledge